test(editor): Move Instance AI runtime coverage (no-changelog)
